Maven Dependency Management and Test Functionality
Now, the API works with Maven Dependency Manager and it is easy to add/remove some packages and there is no need to add manually some extra packages. Also, two sample junit test case is added to ConcertResourceTest.java file in order to help my teammates. 
Also, some other unit test cases will be added to this branch. Please do not merge this branch to the master branch for now.